bugfix: data corruption on GC during reads
sometimes, read-image and read-image-info would return garbage w, h
and channels (as in values like -432522345).

it turned out it was the x, y and channels objects that got moved
around during chicken_port_read and friends, resulting in
stbi_load_from_callbacks writing the final results somewhere
else (bad).

this solution constructs the 4 return values in C instead. load-image
and friends are not affected by this as they are not callbacks.

I came across this as was doing:

    $ cat pic.scm
    (use stb-image stb-image-resize)
    (port-for-each
     (lambda (line)
       (print (handle-exceptions e #f
                  (receive (with-input-from-file line read-image-info)))
              " " line))
     read-line)
    $ find ~/pictures/ -type f | csi -s pic.scm